// angular/src/app/main/supplyChain/sales/shared/dtos/creditDebitNoteDetail-dto.ts
export class CreditDebitNoteDetailDto{
    srNo: number;
    itemId: number;
    description: string;
    unit: string;
    conver: any;
    qty: number;
	rate: number;
	//disc: number;
    amount: number;
    // taxAuth:string;
	// taxClass:number;
	// taxRate:number;
	// taxAmt:number;
    remarks:string;
	//netAmount:number;
}